OneDrive/Рабочий стол/репозиторий/lab 5/shop/requirements.txt
Pillow==8.0.1
Django>2.0
psycopg2-binary>=2.8
django-crispy-forms<=1.9.2
django-cleanup<=5.0.0
